requirements.txt
pydantic-ai-slim[openai,duckduckgo]>=1.73.0
httpx>=0.27.0
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
streamlit>=1.40.0
python-dotenv>=1.0.0
